+++ veltahleon/src/main/java/com/testament/veltahleon/rest/calendar/DayController.java
@@ -163,6 +163,18 @@ public class DayController {
         );
     }
 
+    @PutMapping("/update/day")
+    public ResponseEntity<CustomResponse> updateDay(@RequestBody Day day) {
+        return ResponseEntity.ok(CustomResponse.builder()
+                .timestamp(LocalDateTime.now())
+                .status(HttpStatus.OK)
+                .statusCode(HttpStatus.OK.value())
+                .data(Map.of("dataUpdated", dayService.updateDay(day)))
+                .message("Day updated!")
+                .build()
+        );
+    }
+
 //    @ResponseStatus(HttpStatus.BAD_REQUEST)
 //    @ExceptionHandler(MethodArgumentNotValidException.class)
 //    public ResponseEntity<CustomResponse> handleValidationExceptions(