+++ veltahleon/src/main/java/com/testament/veltahleon/services/ifc/calendar/DayService.java
@@ -19,5 +19,7 @@ public interface DayService {
     Collection<Day> saveDays(Collection<Day> days);
     Day updateDay(Long id, Day day);
 
+    Day updateDay(Day day);
+
     //Day update(Long id, Day day);
 }